backend/requirements.txt
boto3==1.9.125
botocore==1.12.125
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-storages==1.7.1
django-webpack-loader==0.6.0
docutils==0.14
get==2019.3.22
idna==2.8
jmespath==0.9.4
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
post==2019.3.22
public==2019.3.22
python-dateutil==2.8.0
pytz==2018.9
query-string==2019.3.22
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
